LayoutTests/imported/w3c/web-platform-tests/inert/inert-and-contenteditable.tentative-expected.txt


FAIL execCommand should return the same value assert_equals: [false,true,false,false] at index 1 expected false but got true
FAIL The resulting textContent should be the same value assert_equals: ["{target}","{}","{target}","{target}"] at index 1 expected "{target}" but got "{}"
{
target

}
{}
{
target

}
{
target

}
